added test for organization account verification strategy

+++ src/services/Organization/OrganizationAccount/OrganizationAccountVerificationStrategy.ts
@@ -27,15 +27,21 @@ export default class OrganizationAccountVerificationStrategy implements IVerific
 
     async verify() {
         const user = await this.userBroker.findUserByEmail(this.email);
+        if (!user) {
+            throw new Error(`No user with email ${this.email}`)
+        }
         const tokens = await this.tokenBroker.getTokens(user)
         const verificationTokens = tokens.filter((x) => x.scope.includes(Scope.VerifyEmail));
+        if (verificationTokens.length === 0) {
+            throw new Error(`No verification tokens for user with email: ${this.email}`)
+        }
         const values = await this.tokenService.decryptToken(verificationTokens[0]);
         if (values.get("code") === this.verificationCode) {
             await this.tokenService.deleteExistingToken(user)
             user.verified = true;
             return await user.save();
         } else {
-            throw new Error("Failed to verify user, incorrect verification PIN");
+            throw new Error("Failed to verify user, incorrect verification code");
         }
     }
 }
